% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addthresholds.R
\name{addthresholds}
\alias{addthresholds}
\title{Extract Threshold Values}
\usage{
addthresholds(df, qtr = NULL)
}
\arguments{
\item{df}{dataframe (if it contains current quarter, otherwise specify with qtr)}

\item{qtr}{current fiscal quarter, eg "q2", default = NULL}
}
\description{
Extract Threshold Values
}
