% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R
\name{assemble}
\alias{assemble}
\title{Assemble scores by site}
\usage{
assemble(filepath, template_filepath = NULL, output_folderpath = NULL)
}
\arguments{
\item{filepath}{full filepath for MSD site file}

\item{template_filepath}{where is the SIMS template file located?}

\item{output_folderpath}{what folder do you want the output saved to?}
}
\description{
Assemble scores by site
}
